first pass
I rewrote the code from the book solution in a way that I could better understand what was happening at each stage. I modified this to take input text of song lyrics and write a song. I'd like to flesh this out to write verses and a chorus, but need to move on to the stats work. This was fun, though, and I'd love to revisit it. I ran it with song lyrics by Björk and it produced some interesting results!